staticState
The staticState member is designed to be used as an internal memory
of a state which will not mutate as the State Machine instance traverses through the state cycle.

Say you have an application where a button push is supposed to put your device in a certain state.

Despite this, you want the call stack to return to your idle method to perform arbitrary tasks such as polling registers, checking buttons, reading sensors or other tasks.

If nothing changes - no reason to exit the state it was in last - it should keep transitioning to the state it was in before.

the staticState is used for this purpose and eliminates any external variables needed to keep track of the recent most state.